refactor : form과 editor 컴포넌트 분리
