+++ 신동현/src/pages/index.tsx
@@ -6,7 +6,7 @@ import KaKaoForm from "./SignUp/KakaoForm";
 import KakaoLoginLoading from "./SignUp/KakaoLoginLoading";
 import Detail from "./Detail/Detail";
 import Write from "./Write/Write";
-import WriteEditor from "./Write/WriteEditor";
+import WriteFrom from "./Write/WriteForm";
 import Update from "./Update/Update";
 import UpdateForm from "./Update/UpdateForm";
 import Root from "./Root";
@@ -16,4 +16,4 @@ import EmailUpdateForm from "./MyPage/EmailUpdateForm";
 import KaKaoUpdateForm from "./MyPage/KakaoUpdateForm";
 import MyPageDetail from "./MyPage/MyPageDetail";
 
-export { Home, SignUp, SignUpDetail, EmailForm, KaKaoForm, KakaoLoginLoading, Detail, Write, WriteEditor, Update, UpdateForm, Root, NotFound, MyPage, EmailUpdateForm, KaKaoUpdateForm, MyPageDetail };
+export { Home, SignUp, SignUpDetail, EmailForm, KaKaoForm, KakaoLoginLoading, Detail, Write, WriteFrom, Update, UpdateForm, Root, NotFound, MyPage, EmailUpdateForm, KaKaoUpdateForm, MyPageDetail };
